add cron function hello

+++ app/controllers/categories_controller.rb
@@ -12,7 +12,7 @@ class CategoriesController < ApplicationController
 		@posts = Post.where(country_id: @country.id).where.not(user_id: [user, user2]).order(created_at: :desc).includes(:user, :article).limit(30)
 
 		if current_user && current_user.feed_notifications.any?
-			current_user.feed_notifications.delete_all
+			current_user.feed_notifications.where(country_id: current_user.country_id).delete_all
 		end
 	end
 	
@@ -22,7 +22,7 @@ class CategoriesController < ApplicationController
 	
 	def top
 		@mode = cookies[:mode] || "on"
-		@articles = Article.all.where(category_id: Category.all.pluck(:id), country_id: @country.id).where("published_time >= ?", Time.now.ago(2.days)).where.not(posts_count: 0).order(priority_level: :desc, e_indecator: :desc, published_time: :desc).limit(30)
+		@articles = Article.all.where(category_id: Category.all.pluck(:id), country_id: @country.id).where.not(posts_count: 0).order(priority_level: :desc, e_indecator: :desc, published_time: :desc).limit(30)
 		@featured_article = @articles.first
 		unless @featured_article.nil?
 			@best_post = @articles.first.posts.order(likes_count: :desc).includes(:user).limit(1).first
